ai_assistant/requirements.txt
python-dotenv>=1.0.0
openai
langchain
streamlit
python-docx
schedule
rich
